requirements.txt
google-genai
python-dotenv
fastapi[all]
uvicorn[all]
python-multipart
